Add sort_value column on the db
Summary:

Makes it so we can sort items. Initially, we're not going to expose the sort value to bfmodel etc. Eventually we might, but for now the sort value is the js date epoch and only gets generated on insert.

I anticipate adding it to bfmodel when we want to do things like update the sort order of something.

When we use cursors in graphql, if the sort order gets corrupted b/c a sort value gets changed, we'll have to handle that case.

Test Plan: