Enforce "verbatim-module-syntax"
Summary:

If importing an item just for its type, annotating it with "type" will ensure either the individual or entire import is stripped before being compiled by esbuild / typescript.

Luckily there was a codemod to run over the entire codebase, which is the majority of files in this commit.

Test Plan:

```
deno lint
```